ui/table: Extract the TableRow outside of VolumeListTable
Refs: #2931
