+++ ui/src/components/VolumeListTable.js
@@ -22,6 +22,7 @@ import {
 import CircleStatus from './CircleStatus';
 import { Button, ProgressBar, Tooltip } from '@scality/core-ui';
 import { intl } from '../translations/IntlGlobalProvider';
+import TableRow from './TableRow';
 import {
   VOLUME_CONDITION_LINK,
   VOLUME_CONDITION_UNLINK,
@@ -74,24 +75,6 @@ const VolumeListContainer = styled.div`
   .sc-progressbarcontainer > div {
     background-color: ${(props) => props.theme.brand.secondaryDark1};
   }
-`;
-
-const TableRowStyle = styled.div`
-  &:hover,
-  &:focus {
-    background-color: ${(props) => props.theme.brand.backgroundBluer};
-    outline: none;
-    cursor: pointer;
-  }
-
-  &:last-child {
-    border: none;
-  }
-
-  background-color: ${(props) =>
-    props.volumeName === props.row.values.name
-      ? props.theme.brand.backgroundBluer
-      : props.theme.brand.primary};
 `;
 
 // * table body
@@ -175,66 +158,6 @@ function GlobalFilter({
     />
   );
 }
-
-const TableRow = (props) => {
-  const { row, style, rowClicked, volumeName, theme } = props;
-
-  return (
-    <TableRowStyle
-      {...row.getRowProps({
-        onClick: () => rowClicked(row),
-        // Note:
-        // We need to pass the style property to the row component.
-        // Otherwise when we scroll down, the next rows are flashing because they are re-rendered in loop.
-        style: { ...style, marginLeft: '5px' },
-      })}
-      volumeName={volumeName}
-      row={row}
-    >
-      {row.cells.map((cell) => {
-        let cellProps = cell.getCellProps({
-          style: {
-            ...cell.column.cellStyle,
-            // Vertically center the text in cells.
-            display: 'flex',
-            flexDirection: 'column',
-            justifyContent: 'center',
-          },
-        });
-
-        if (cell.column.Header === 'Name') {
-          return (
-            <div {...cellProps} data-cy="volume_table_name_cell" className="td">
-              {cell.render('Cell')}
-            </div>
-          );
-        } else if (cell.column.Header !== 'Name' && cell.value === undefined) {
-          return (
-            <div {...cellProps} className="td">
-              <Tooltip
-                placement="top"
-                overlay={
-                  <TooltipContent>{intl.translate('unknown')}</TooltipContent>
-                }
-              >
-                <UnknownIcon
-                  className="fas fa-minus"
-                  theme={theme}
-                ></UnknownIcon>
-              </Tooltip>
-            </div>
-          );
-        } else {
-          return (
-            <div {...cellProps} className="td">
-              {cell.render('Cell')}
-            </div>
-          );
-        }
-      })}
-    </TableRowStyle>
-  );
-};
 
 function Table({
   columns,
